pyproject.toml
[project]
name = "hapax-sdlc"
version = "0.1.0"
description = "Shared SDLC pipeline infrastructure for hapax repositories"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.12",
]

[project.optional-dependencies]
ci = []

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
[tool.ruff.lint.isort]
known-first-party = ["sdlc", "shared", "scripts"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
